Extract a Pokédex data file
