+++ src/_pokedex.js
@@ -1,0 +1,24 @@
+const pokedex = {
+    miles: {
+        key: 'miles',
+        name: 'Miles'
+    },
+    miles_evolved: {
+        key: 'miles',
+        name: 'Kilometers'
+    },
+    holli: {
+        key: 'holli',
+        name: 'Holli'
+    },
+    ian: {
+        key: 'ian',
+        name: 'Ian'
+    },
+    ainsley: {
+        key: 'ainsley',
+        name: 'Ainsley'
+    }
+};
+
+export default pokedex;
